[FSSDK-10941] move notification_center to the lib root directory
