+++ lib/core/decision_service/index.tests.js
@@ -26,7 +26,7 @@ import {
 } from '../../utils/enums';
 import { createLogger } from '../../plugins/logger';
 import { getForwardingEventProcessor } from '../../event_processor/forwarding_event_processor';
-import { createNotificationCenter } from '../notification_center';
+import { createNotificationCenter } from '../../notification_center';
 import Optimizely from '../../optimizely';
 import OptimizelyUserContext from '../../optimizely_user_context';
 import projectConfig, { createProjectConfig } from '../../project_config/project_config';
