+++ lib/notification_center/index.tests.js
@@ -17,9 +17,9 @@ import sinon from 'sinon';
 import { assert } from 'chai';
 
 import { createNotificationCenter } from './';
-import * as enums from '../../utils/enums';
-import { createLogger } from '../../plugins/logger';
-import errorHandler from '../../plugins/error_handler';
+import * as enums from '../utils/enums';
+import { createLogger } from '../plugins/logger';
+import errorHandler from '../plugins/error_handler';
 
 var LOG_LEVEL = enums.LOG_LEVEL;
