+++ lib/optimizely/index.tests.js
@@ -34,7 +34,7 @@ import * as jsonSchemaValidator from '../utils/json_schema_validator';
 import * as projectConfig from '../project_config/project_config';
 import testData from '../tests/test_data';
 import { getForwardingEventProcessor } from '../event_processor/forwarding_event_processor';
-import { createNotificationCenter } from '../core/notification_center';
+import { createNotificationCenter } from '../notification_center';
 import { createProjectConfig } from '../project_config/project_config';
 import { getMockProjectConfigManager } from '../tests/mock/mock_project_config_manager';
