+++ lib/optimizely_user_context/index.tests.js
@@ -23,7 +23,7 @@ import { NOTIFICATION_TYPES } from '../utils/enums';
 
 import OptimizelyUserContext from './';
 import { createLogger } from '../plugins/logger';
-import { createNotificationCenter } from '../core/notification_center';
+import { createNotificationCenter } from '../notification_center';
 import Optimizely from '../optimizely';
 import errorHandler from '../plugins/error_handler';
 import { CONTROL_ATTRIBUTES, LOG_LEVEL, LOG_MESSAGES } from '../utils/enums';
